fix silly cast bug
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@9797 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/wp/ap/xp/ap_EditMethods.cpp
@@ -10190,7 +10190,7 @@ Defun(resizeImage)
 		}
 
 		UT_Rect r = orgImgRect;
-		double aRatio = r.height/r.width;
+		double aRatio = ((double )r.height)/((double )r.width);
 		
 		// we can savely use the cursor format to see what kind of dragging we are doing
 		GR_Graphics::Cursor cur = pView->getImageSelCursor();
